// app/types/works.ts
type Works = {
  id: number
  title: string
  categories: {
    nodes: {
      name: string
    }[]
  }
  date: string
  featuredImage: any
  node: {
    id: number
    title: string
    content: string
    date: string
    featuredImage: {
      node: {
        altText: string
        mediaDetails: {
          height: number
          width: number
        }
        sourceUrl: string
      }
    }
    slug: string
  }
  slug: string
}

export default Works